added first tests for the core module as well as expsing access to attributes
